다른테이블값 불러오기 완료

+++ src/main/java/com/artfactory/project01/todayart/service/BoardCategoryService.java
@@ -3,9 +3,11 @@ package com.artfactory.project01.todayart.service;
 import com.artfactory.project01.todayart.entity.BoardCategory;
 import com.artfactory.project01.todayart.model.BoardCategoryForm;
 import com.artfactory.project01.todayart.repository.BoardCategoryRepository;
+import com.artfactory.project01.todayart.util.PrincipalUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,6 +17,7 @@ public class BoardCategoryService {
 
     @Autowired
     private BoardCategoryRepository boardCategoryRepository;
+
 
     /*
        작성자: 진표
@@ -31,7 +34,7 @@ public class BoardCategoryService {
        작성자: 진표
        기능 : 보드 카테고리 리스트
        @param 보드ID(Integer)
-       @return Page<Article>
+       @return Page<BoardCategory>
     */
     @Transactional(readOnly = true)
     public Page<BoardCategory> listOfBoardCategory(Pageable pageable) {
@@ -42,7 +45,7 @@ public class BoardCategoryService {
        작성자: 진표
        기능 : 보드 카테고리 업데이트
        @param BoardCategoryForm
-       @return 업데이트 된 Article
+       @return 업데이트 된 BoardCategoryForm
     */
     @Transactional
     public BoardCategory updateBoardCategory(Integer id, BoardCategoryForm boardCategoryForm) {
@@ -59,9 +62,11 @@ public class BoardCategoryService {
        @param ArticleId
     */
     @Transactional
-    public void deleteBoardCartegory(Integer id) {
-        boardCategoryRepository.deleteById(id);
+    public void deleteBoardCartegory(Integer id) throws Exception {
+        if(id ==1 || id ==2) {
+            throw new Exception("기본 게시판은 삭제할 수 없습니다");
+        } else{
+            boardCategoryRepository.deleteById(id);
+        }
     }
-
-
 }
